Move to integration tests
